change the type of file_size and file_offset to off_t

+++ common/shared_func.c
@@ -820,7 +820,7 @@ int init_pthread_attr(pthread_attr_t *pattr)
 	return 0;
 }
 
-int getFileContent(const char *filename, char **buff, int *file_size)
+int getFileContent(const char *filename, char **buff, off_t *file_size)
 {
 	int fd;
